+++ .idea/misc.xml
@@ -57,6 +57,7 @@
         <entry key="app/src/main/res/drawable/circular_view.xml" value="0.22037037037037038" />
         <entry key="app/src/main/res/layout/action_bar.xml" value="0.29791666666666666" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.29791666666666666" />
+        <entry key="app/src/main/res/layout/dialog_text_entry.xml" value="0.29791666666666666" />
         <entry key="app/src/main/res/layout/fragment_add_mood_entry.xml" value="0.29791666666666666" />
         <entry key="app/src/main/res/layout/fragment_add_task_entry.xml" value="0.29791666666666666" />
         <entry key="app/src/main/res/layout/fragment_calendar.xml" value="0.159375" />
@@ -64,6 +65,11 @@
         <entry key="app/src/main/res/layout/fragment_home_bottom_navigation.xml" value="0.25" />
         <entry key="app/src/main/res/layout/fragment_more.xml" value="0.29791666666666666" />
         <entry key="app/src/main/res/layout/fragment_upsert_mood_task_icon.xml" value="0.29791666666666666" />
+        <entry key="app/src/main/res/layout/item_icon_list.xml" value="0.29791666666666666" />
+        <entry key="app/src/main/res/layout/item_rv_empy_state.xml" value="0.29791666666666666" />
+        <entry key="app/src/main/res/layout/item_rv_single_selection.xml" value="0.29791666666666666" />
+        <entry key="app/src/main/res/layout/item_rv_task_icon.xml" value="0.29791666666666666" />
+        <entry key="app/src/main/res/layout/item_text_chips.xml" value="0.29791666666666666" />
         <entry key="app/src/main/res/layout/partial_layout_profile_section.xml" value="0.25" />
         <entry key="app/src/main/res/layout/task_entry_action_bar.xml" value="0.29791666666666666" />
         <entry key="app/src/main/res/menu/menu_bottom_navigation.xml" value="0.29791666666666666" />
@@ -72,6 +78,8 @@
         <entry key="core/presentation/src/main/res/drawable/bg_rounded_white_ripple.xml" value="0.1925925925925926" />
         <entry key="core/presentation/src/main/res/drawable/bg_rounded_white_stroke_gray.xml" value="0.1925925925925926" />
         <entry key="core/presentation/src/main/res/drawable/bg_rounded_white_stroke_gray_ripple.xml" value="0.1925925925925926" />
+        <entry key="core/presentation/src/main/res/layout/dialog_text_entry.xml" value="0.29791666666666666" />
+        <entry key="core/presentation/src/main/res/layout/task_entry_action_bar.xml" value="0.29791666666666666" />
       </map>
     </option>
   </component>
