+++ app/src/main/java/lostankit7/droid/moodtracker/presentation/fragment/edit/editmood/DisplayMoodIconsFragment.kt
@@ -5,10 +5,10 @@ import android.view.MenuItem
 import androidx.lifecycle.ViewModelProvider
 import lostankit7.droid.moodtracker.R
 import lostankit7.droid.moodtracker.base.fragment.BaseDaggerFragment
+import lostankit7.droid.moodtracker.core_presentation.databinding.CommonActionBarBinding
 import lostankit7.droid.moodtracker.data.database.entities.Icon
 import lostankit7.droid.moodtracker.data.database.entities.MoodIcon
 import lostankit7.droid.moodtracker.databinding.FragmentDisplayListBinding
-import lostankit7.droid.moodtracker.databinding.TaskEntryActionBarBinding
 import lostankit7.droid.moodtracker.di.AppComponent
 import lostankit7.droid.moodtracker.presentation.adapter.IconListRvAdapter
 import lostankit7.droid.moodtracker.presentation.viewmodel.MoodEntryViewModel
@@ -49,7 +49,7 @@ class DisplayMoodIconsFragment :
         }
     }
 
-    override fun updateActionBar(actionBar: TaskEntryActionBarBinding) = with(actionBar) {
+    override fun updateActionBar(actionBar: CommonActionBarBinding) = with(actionBar) {
         super.updateActionBar(actionBar)
         showBackButton()
     }
