+++ app/src/main/java/lostankit7/droid/moodtracker/presentation/fragment/edit/editmood/UpsertMoodIconFragment.kt
@@ -7,15 +7,15 @@ import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import lostankit7.droid.moodtracker.base.fragment.BaseDaggerFragment
+import lostankit7.droid.moodtracker.core_presentation.databinding.CommonActionBarBinding
 import lostankit7.droid.moodtracker.data.database.entities.Icon
 import lostankit7.droid.moodtracker.data.database.entities.Suggestion
 import lostankit7.droid.moodtracker.databinding.FragmentUpsertMoodTaskIconBinding
-import lostankit7.droid.moodtracker.databinding.TaskEntryActionBarBinding
 import lostankit7.droid.moodtracker.di.AppComponent
 import lostankit7.droid.moodtracker.presentation.adapter.MoodIconRvAdapter
 import lostankit7.droid.moodtracker.presentation.adapter.TextRvAdapter
 import lostankit7.droid.moodtracker.presentation.viewmodel.MoodEntryViewModel
-import lostankit7.droid.moodtracker.utils.hideKeyBoard
+import lostankit7.droid.moodtracker.core_presentation.utils.hideKeyBoard
 import lostankit7.droid.moodtracker.utils.showBackAndSaveButton
 
 class UpsertMoodIconFragment :
@@ -74,7 +74,7 @@ class UpsertMoodIconFragment :
         }
     }
 
-    override fun updateActionBar(actionBar: TaskEntryActionBarBinding) = with(actionBar) {
+    override fun updateActionBar(actionBar: CommonActionBarBinding) = with(actionBar) {
         super.updateActionBar(actionBar)
         showBackAndSaveButton()
     }
