+++ core/presentation/src/main/res/layout/common_action_bar.xml
