+++ core/presentation/src/main/res/layout/item_text_chips.xml
