+++ core/presentation/src/main/res/values/strings.xml
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
+    <string name="text_save">Save</string>
+    <string name="text_add">Add</string>
+    <string name="text_cancel">Cancel</string>
+    <string name="text_edit">Edit</string>
+    <string name="text_delete">Delete</string>
 </resources>
